# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentAssignmentPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "assignee_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee User Id"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "assignee_user_id",
          "expected_version"
        ],
        "title": "IncidentAssignmentPatchRequest",
        "type": "object"
      },
      "IncidentAuditListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentAuditResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "IncidentAuditListResponse",
        "type": "object"
      },
...
      "IncidentDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/IncidentContextResponse"
          },
          "evidence_memberships": {
            "items": {
              "$ref": "#/components/schemas/IncidentMembershipResponse"
            },
            "title": "Evidence Memberships",
            "type": "array"
          },
          "incident": {
            "$ref": "#/components/schemas/IncidentRecordResponse"
          },
          "lineage_references": {
            "items": {
              "$ref": "#/components/schemas/IncidentLineageReference"
            },
            "title": "Lineage References",
            "type": "array"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/IncidentNoteResponse"
            },
            "title": "Notes",
            "type": "array"
          },
          "severity_history": {
            "items": {
              "$ref": "#/components/schemas/IncidentSeverityHistoryResponse"
            },
            "title": "Severity History",
            "type": "array"
          },
          "status_history": {
            "items": {
              "$ref": "#/components/schemas/IncidentStatusHistoryResponse"
            },
            "title": "Status History",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/IncidentTimelineResponse"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "incident",
          "evidence_memberships",
          "lineage_references",
          "timeline",
          "status_history",
          "severity_history",
          "notes",
          "context"
        ],
        "title": "IncidentDetailResponse",
        "type": "object"
      },
...
      "IncidentListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentRecordResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "limit",
          "next_cursor"
        ],
        "title": "IncidentListResponse",
        "type": "object"
      },
...
      "IncidentMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "incident": {
            "$ref": "#/components/schemas/IncidentRecordResponse"
          },
          "operation": {
            "enum": [
              "status_changed",
              "note_added",
              "assignment_changed",
              "disposition_changed"
            ],
            "title": "Operation",
            "type": "string"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "incident",
          "operation"
        ],
        "title": "IncidentMutationResponse",
        "type": "object"
      },
      "IncidentNoteCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "content",
          "expected_version"
        ],
        "title": "IncidentNoteCreateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OTSOCSessionCookie": {
        "description": "Opaque HttpOnly local session cookie. The configured runtime name may differ.",
        "in": "cookie",
        "name": "otsoc_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Repository and contract foundation for the isolated, advisory-only OT-SOC Fusion X academic prototype.",
    "title": "OT-SOC Fusion X",
    "version": "1.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "read_incident_list_api_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "CURRENT",
              "enum": [
                "CURRENT",
                "ALL_HISTORY",
                "RUN"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "in": "query",
            "name": "observed_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observed From"
            }
          },
          {
            "in": "query",
            "name": "observed_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observed To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Incident List",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "read_incident_detail_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Incident Detail",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/assignment": {
      "patch": {
        "operationId": "patch_incident_assignment_api_v1_incidents__incident_id__assignment_patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentAssignmentPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Patch Incident Assignment",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/audit": {
      "get": {
        "operationId": "read_incident_audit_api_v1_incidents__incident_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentAuditListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Incident Audit",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/disposition": {
      "patch": {
        "operationId": "patch_incident_disposition_api_v1_incidents__incident_id__disposition_patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentDispositionPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Patch Incident Disposition",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/notes": {
      "post": {
        "operationId": "create_incident_note_api_v1_incidents__incident_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentNoteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Create Incident Note",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/report": {
      "get": {
        "operationId": "read_incident_report_api_v1_incidents__incident_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Incident Report",
        "tags": [
          "incidents"
        ]
      },
      "put": {
        "operationId": "put_incident_report_api_v1_incidents__incident_id__report_put",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentReportPutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Put Incident Report",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/status": {
      "patch": {
        "operationId": "patch_incident_status_api_v1_incidents__incident_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentStatusPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Patch Incident Status",
        "tags": [
          "incidents"
        ]
      }
    },
...
  }
}